+++ red_bot/settings/setting.py
@@ -30,10 +30,11 @@ dp = Dispatcher(
 )
 
 def main():
-    from red_bot import app
 
+    from red_bot import app
 
     executor.start_polling(
         dp,
         skip_updates = True
     )
+
