+++ red_bot/sql_db/bot_tables.py
@@ -3,7 +3,8 @@ import logging
 import sqlite3
 
 
-from red_bot.settings import config
+from red_bot.settings.config import DATA_PATH
+from red_bot.settings.config import DB_NAME
 
 
 class Bot_tables_DB:
@@ -54,14 +55,6 @@ class Bot_tables_DB:
         self.conn.commit()
         logging.info('--- Table "POSTS" has been created ---')
 
-    def drop_posts_table(self) -> None:
-        self.cur.execute(
-            '''
-            DROP TABLE posts;
-            '''
-        )
-        self.conn.commit()
-
     def create_responders_tables(self) -> None:
         self.conn
         self.cur.execute(
@@ -77,19 +70,10 @@ class Bot_tables_DB:
         self.conn.commit()
         logging.info('--- Table "RESPONDERS" has been created ---')
 
-    def drop_responders_table(self) -> None:
-        self.cur.execute(
-            '''
-            DROP TABLE responders;
-            '''
-        )
-        self.conn.commit()
-
-
 def create_table() -> None:
     DB_tables = Bot_tables_DB(
-        name = config.DB_NAME,
-        path = config.DB_PATH
+        name = DB_NAME,
+        path = DATA_PATH
     )
     DB_tables.create_users_tables()
     DB_tables.create_posts_tables()
@@ -97,7 +81,7 @@ def create_table() -> None:
     logging.info('--- Database for "SEVASTOPOL ADJUTOR BOT" has been created ---')
 
 
-if config.DB_NAME + '.db' not in os.listdir(config.DB_PATH):
+if DB_NAME + '.db' not in os.listdir(DATA_PATH):
     create_table()
 else:
     logging.info('--- Database for "SEVASTOPOL ADJUTOR BOT" connection established ---')
